Flush all accumulations on program termination
All open windows are immediately flushed with their current values,
using the current time as a timestamp instead of the window end time.

This avoids losing stats when the program is restarted, and becomes
necessary for rebalancing which server processes which paths when
Cassabon servers join or leave the group.